[modify] 1. Adjuests backdoor I2C protocol for accessing temperature derating calibration datas from flash.
[add]
1. Adds a call back function to control 0x00.INT_TERR, when temperature derating module is at INIT.
2. Adds a call back function to get BL temperature, when temperature derating module is at INIT.
3. Adds a call back function to access flash for calibration datas, when temperature derating module is at INIT.
4. Adds a switch for temperature derating module.

Date:2022/08/15
SW:T.01.02.07
Project:Ford U717 & CX430

+++ ford_u625/app/module/c_communication_using/M_FPNCtrl.c
@@ -63,7 +63,7 @@ static uint8_t M_FPNCtrl_DeliveryFPNWriteToEE_Process(void)
             if(u8CompResulat == FPN_ENABLE)
             {
                 pu8Delivery_assembly = tFPNCtrlDelivery.pFPNDataBuff;
-                MFixedFlashAccess_WritePage((uint32_t)ADDR_DELIVERY_ASSEMBLY, pu8Delivery_assembly, tFPNCtrlDelivery.FPNDataLength);
+                MFixedFlashAccess_WritePage(ADDR_DELIVERY_ASSEMBLY, pu8Delivery_assembly, tFPNCtrlDelivery.FPNDataLength);
                 u8ProcessStatus = FPN_WRITE_SUCCESS;
             }
             else
@@ -152,7 +152,7 @@ uint8_t M_FPNCtrl_SerNumFPNWriteToEE_Process(void)
             if(u8CompResulat == FPN_ENABLE)
             {
                 pu8SerNum_assembly = tFPNCtrlSerNum.pFPNDataBuff;
-                MFixedFlashAccess_WritePage((uint32_t)ADDR_SERIAL_NUMBER, pu8SerNum_assembly, tFPNCtrlSerNum.FPNDataLength);              
+                MFixedFlashAccess_WritePage(ADDR_SERIAL_NUMBER, pu8SerNum_assembly, tFPNCtrlSerNum.FPNDataLength);              
                 u8ProcessStatus = FPN_WRITE_SUCCESS;
             }
             else
@@ -229,7 +229,7 @@ uint8_t M_FPNCtrl_ProductionByte_FPNWriteToEE_Process(void)
         if(tFPNProductionByte.FPN_Flag_t.FPNDataFlag == FPN_ENABLE)
         {
             pu8FCE = tFPNProductionByte.pFPNDataBuff;
-            MFixedFlashAccess_WritePage((uint32_t)ADDR_PRODUCTION_PHASE_BYTE, pu8FCE, tFPNProductionByte.FPNDataLength);
+            MFixedFlashAccess_WritePage(ADDR_PRODUCTION_PHASE_BYTE, pu8FCE, tFPNProductionByte.FPNDataLength);
             u8ProcessStatus = FPN_WRITE_SUCCESS;
         }
 		else
